plex-live-scan-agent/app/requirements.txt
watchdog==4.0.1
requests==2.32.3
pyyaml==6.0.2
